Less stringent test a bit
